Msutils.Process (mscmp_syst_utils v0.1.0)

This module provides utilities for working with processes.

Functions

register(name)

@spec register(Msutils.Types.Process.name()) ::
  :ok | {:error, Mserror.ProcessUtilsError.t()}

Registers the current process under the given name.

This function must be called from within the process that is to be registered. Attempting to register a process from outside that process will fail.

Parameters

name - The name under which to register the current process.

Returns

:ok - The process was successfully registered.

{:error, %Mserror.ProcessUtilsError{cause: :registration_failed}} - If the registration failed.

{:error, %Mserror.ProcessUtilsError{cause: :invalid_name}} - If the provided name does not comply with Msutils.Types.Process.name/0 naming conventions.

Examples

Register a process locally:

iex> {:ok, pid} = Agent.start_link(fn ->
...>   Msutils.Process.register(:my_process)
...>   %{}
...> end)
iex> {:ok, ^pid} = Msutils.Process.whereis(:my_process)

Register a process globally:

iex> {:ok, pid} = Agent.start_link(fn ->
...>   Msutils.Process.register({:global, :my_global_process})
...>   %{}
...> end)
iex> {:ok, ^pid} = Msutils.Process.whereis({:global, :my_global_process})

Register a process via a registry:

iex> {:ok, pid} = Agent.start_link(fn ->
...>   Msutils.Process.register({:via, Registry, {MscmpSystUtils.TestRegistry, :my_process}})
...>   %{}
...> end)
iex> {:ok, ^pid} = Msutils.Process.whereis({:via, Registry, {MscmpSystUtils.TestRegistry, :my_process}})

Attempt to register with an already taken name:

iex> {:ok, pid1} = Agent.start_link(fn ->
...>   Msutils.Process.register(:taken_name)
...>   %{}
...> end)
iex> {:ok, ^pid1} = Msutils.Process.whereis(:taken_name)
iex> {:ok, _} = Agent.start_link(fn ->
...>   {:error, error} = Msutils.Process.register(:taken_name)
...>   %Mserror.ProcessUtilsError{kind: :registration, cause: :registration_failed} = error
...>   %{}
...> end)

register(registry, name)

@spec register(Msutils.Types.Process.registry(), term()) ::
  :ok | {:error, Mserror.ProcessUtilsError.t()}

Registers the current process under the given registry and name.

This function must be called from within the process that is to be registered. Attempting to register a process from outside that process will fail.

This function provides a convenient way to register processes when the registry and name are known separately at runtime.

Parameters

registry - The registry to use for registration.

name - The name under which to register the current process.

Returns

:ok - The process was successfully registered.

{:error, %Mserror.ProcessUtilsError{cause: :registration_failed}} - If the registration failed.

{:error, %Mserror.ProcessUtilsError{cause: :invalid_name}} - If the provided registry or name combination is invalid.

Examples

Register a process locally:

iex> {:ok, pid} = Agent.start_link(fn ->
...>   Msutils.Process.register(:local, :my_process)
...>   %{}
...> end)
iex> {:ok, ^pid} = Msutils.Process.whereis(:local, :my_process)

Register a process globally:

iex> {:ok, pid} = Agent.start_link(fn ->
...>   Msutils.Process.register(:global, :my_global_process)
...>   %{}
...> end)
iex> {:ok, ^pid} = Msutils.Process.whereis(:global, :my_global_process)

Register a process via a registry:

iex> {:ok, pid} = Agent.start_link(fn ->
...>   Msutils.Process.register({Registry, MscmpSystUtils.TestRegistry}, :my_process)
...>   %{}
...> end)
iex> {:ok, ^pid} = Msutils.Process.whereis({Registry, MscmpSystUtils.TestRegistry}, :my_process)

whereis(name)

@spec whereis(Msutils.Types.Process.name()) ::
  {:ok, pid()} | {:error, Mserror.ProcessUtilsError.t()}

Looks up a process by its name and returns its PID.

Parameters

name - The name of the process whose PID is to be retrieved.

Returns

{:ok, pid} - The PID of the process if found.

{:error, %Mserror.ProcessUtilsError{cause: :process_not_found}} - If the process doesn't exist

{:error, %Mserror.ProcessUtilsError{cause: :invalid_name}} - If the provided name does not comply with Msutils.Types.Process.name/0 naming conventions.

Examples

A simple local process name lookup:

iex> {:ok, pid} = Agent.start_link(fn -> %{} end, name: :example_local_name)
iex> {:ok, ^pid} = Msutils.Process.whereis(:example_local_name)

A global process name lookup:

iex> {:ok, global_pid} = Agent.start_link(fn -> %{} end, name: {:global, :example_global_process})
iex> {:ok, ^global_pid} = Msutils.Process.whereis({:global, :example_global_process})

A via process name lookup:

iex> {:ok, via_pid} = Agent.start_link(fn -> %{} end, name: {:via, Registry, {MscmpSystUtils.TestRegistry, :example_via_process}})
iex> {:ok, ^via_pid} = Msutils.Process.whereis({:via, Registry, {MscmpSystUtils.TestRegistry, :example_via_process}})

Attempted lookup of a non-existent process:

iex> {:error, error} = Msutils.Process.whereis(:nonexistent_process)
iex> %Mserror.ProcessUtilsError{kind: :lookup, cause: :process_not_found} = error

Attempted lookup of an invalid or unsupported process name:

iex> {:error, error} = Msutils.Process.whereis({:invalid, :name})
iex> %Mserror.ProcessUtilsError{kind: :lookup, cause: :invalid_name} = error

whereis(registry, name)

@spec whereis(Msutils.Types.Process.registry(), term()) ::
  {:ok, pid()} | {:error, Mserror.ProcessUtilsError.t()}

Looks up a process by its name and returns its PID.

Typically whereis/2 is called when neither the registry name (or kind) or process name are known until runtime. Since Msutils.Types.Process.name/0, and likewise GenServer.name/0, values vary in construction based on the specific kind of registry, this function provides a convenient way to not have to worry about the name construction. If the registry kind is known at compile time, then whereis/1 is preferred as it should be slightly more efficient.

Parameters

registry - The registry to use for the lookup.

name - The name of the process whose PID is to be retrieved.

Returns

{:ok, pid} - The PID of the process if found.

{:error, %Mserror.ProcessUtilsError{cause: :process_not_found}} - If the process doesn't exist

{:error, %Mserror.ProcessUtilsError{cause: :invalid_name}} - If the provided name does not comply with Msutils.Types.Process.name/0 naming conventions.

Examples

A simple local process name lookup:

iex> {:ok, pid} = Agent.start_link(fn -> %{} end, name: :example2_local_name)
iex> {:ok, ^pid} = Msutils.Process.whereis(:local, :example2_local_name)

A global process name lookup:

iex> {:ok, global_pid} = Agent.start_link(fn -> %{} end, name: {:global, :example2_global_process})
iex> {:ok, ^global_pid} = Msutils.Process.whereis(:global, :example2_global_process)

A via process name lookup:

iex> {:ok, via_pid} = Agent.start_link(fn -> %{} end, name: {:via, Registry, {MscmpSystUtils.TestRegistry, :example2_via_process}})
iex> {:ok, ^via_pid} = Msutils.Process.whereis({Registry, MscmpSystUtils.TestRegistry}, :example2_via_process)

Attempted lookup of a non-existent process:

iex> {:error, error} = Msutils.Process.whereis(:local, :nonexistent_process)
iex> %Mserror.ProcessUtilsError{kind: :lookup, cause: :process_not_found} = error

Attempted lookup of an invalid or unsupported process name:

iex> {:error, error} = Msutils.Process.whereis(:local, {:invalid, :name})
iex> %Mserror.ProcessUtilsError{kind: :lookup, cause: :invalid_name} = error