Msutils.Types.Process (mscmp_syst_utils v0.1.0)

This module defines process related types used by the MscmpSystUtils library.

Types

name()

@type name() :: pid() | atom() | {:global, term()} | {:via, module(), term()}

The registry qualified name of a process for use in pid/0 lookups.

This type is compatible with the GenServer.name/0 type with the addition of pid/0 as a valid name.

registry()

@type registry() :: :local | :global | {module(), term()}

The registry to use when either looking up a process pid/0 or registering a process's name.

The currently supported registries are:

:local - The local registry. Service names must be atoms.

:global - The global registry.

{module(), term()} - A registry module and name. The registry must to usable in :via tuple names. See the GenServer documentation regarding name registration for details and requirements.